requirements.txt
click==8.1.7
fastapi==0.108.0
python-dotenv==1.0.0
starlette==0.32.0.post1
uvicorn==0.25.0
